Add back code that went missing.
I'm not sure why I removed this.  I probably did it while making warnings go away.  However, it was stupid, and was the root of why I was so confused trying to understand the outbound audio code path.  I had removed the line of code responsible for filling in the buffer with audio from this class.  Oops.  :-(